+++ doxray.js
@@ -44,7 +44,7 @@ Doxray.prototype.run = function( src, options, callback ) {
   var parsed;
   src = require('./utils.js').handleSrc( src );
   options = require('./utils.js').handleOptions( options );
-  parsed = this.parse( src, options.merge );
+  parsed = this.parse( src, options );
   if ( options.processors ) {
     parsed = this.postParseProcessing( parsed, options.processors );
   }
@@ -57,15 +57,15 @@ Doxray.prototype.run = function( src, options, callback ) {
   return parsed;
 };
 
-Doxray.prototype.parse = function( src, merge ) {
+Doxray.prototype.parse = function( src, options ) {
   var parsed = [];
   src.forEach(function( singleSrc ) {
     var fileContents, docs, code, convertedDocs, ext;
     // Get the file extension for src so we know which regex to use.
     ext = require('./utils.js').getCommentType( singleSrc );
-    fileContents = require('./utils.js').getFileContents( singleSrc, this.regex[ ext ] );
-    docs = require('./utils.js').parseOutDocs( fileContents, this.regex[ ext ] );
-    code = require('./utils.js').parseOutCode( fileContents, this.regex[ ext ] );
+    fileContents = require('./utils.js').getFileContents( singleSrc, options.regex[ ext ] );
+    docs = require('./utils.js').parseOutDocs( fileContents, options.regex[ ext ] );
+    code = require('./utils.js').parseOutCode( fileContents, options.regex[ ext ] );
     if ( docs.length === 0 ) {
       convertedDocs = [];
     } else {
@@ -115,40 +115,3 @@ Doxray.prototype.writeJS = function( convertedDocs, dest, callback ) {
 };
 
 module.exports = Doxray;
-
-// Doxray.prototype.mergeParsedSources = function( sources ) {
-//   var equal, first, theRest, nonMatches;
-//   equal = require('deep-equal');
-//   // The first parsed source will be our "master" source.
-//   first = sources[ 0 ];
-//   // The rest of the sources are saved to their own array.
-//   theRest = sources.slice( 1 );
-//   // We'll need this later for doc sets that don't match anything.
-//   uniqueSets = [];
-//   // Compare each doc set from the master source to the doc sets of the rest of
-//   // the sources. If both sets of docs are exactly the same then merge them
-//   // together. "Merging" is done by taking the code from the second doc set and
-//   // adding to the code array of the master doc set.
-//   theRest.forEach( function( src ) {
-//     src.forEach( function( secondDocSet ) {
-//       var matchesSomethingInFirst = false;
-//       first.forEach( function( firstDocSet ) {
-//         if ( equal( firstDocSet.docs, secondDocSet.docs ) ) {
-//           // If this doc set has docs that match another doc sets docs,
-//           // then take only its code and move it to the matching set.
-//           firstDocSet.code = firstDocSet.code.concat( secondDocSet.code );
-//           matchesSomethingInFirst = true;
-//         }
-//       }, this );
-//       if ( matchesSomethingInFirst === false ) {
-//         // If this doc set has docs that do not match any other doc sets docs,
-//         // then move this whole doc set into uniqueSets. uniqueSets will be
-//         // added to first after we finish looping through it.
-//         uniqueSets.push( secondDocSet );
-//       }
-//     }, this );
-//   }, this );
-//   // Add all the unique doc sets that couldn't get merged.
-//   first = first.concat( uniqueSets );
-//   return first;
-// };